Move all burrito types to central file
