+++ src/renderer/src/burrito/BurritoWrapper.tsx
@@ -4,7 +4,6 @@ import StickyRedirect from '../components/StickyRedirect';
 import { useOrgDefaults } from '../crud/useOrgDefaults';
 import { BurritoHeader } from '../components/BurritoHeader';
 import { MetadataView } from './MetadataView';
-import type { BurritoWrapper as BurritoWrapperType } from './data/wrapperBuilder';
 import { wrapperBuilder } from './data/wrapperBuilder';
 import { useOrbitData } from '../hoc/useOrbitData';
 import {
@@ -18,12 +17,12 @@ import {
 import related from '../crud/related';
 import { burritoContents } from './BurritoContents';
 import { BurritoType } from './BurritoType';
-import { Burrito } from './data/wrapperBuilder';
 import { AltButton } from '../control/AltButton';
 import packageJson from '../../package.json';
 import { toCamel } from '../utils';
 import { shallowEqual, useSelector } from 'react-redux';
 import { burritoSelector, sharedSelector } from '../selector';
+import { BurritoEntry, BurritoWrapper } from 'burrito/data/types';
 const version = packageJson.version;
 const productName = packageJson.build.productName;
 
@@ -38,7 +37,7 @@ export function BurritoWrapper() {
   const bibles = useOrbitData<BibleD[]>('bible');
   const [view, setView] = React.useState('');
   const [refresh, setRefresh] = React.useState(0);
-  const [metaData, setMetaData] = React.useState<BurritoWrapperType>();
+  const [metaData, setMetaData] = React.useState<BurritoWrapper>();
   const { getOrgDefault, setOrgDefault } = useOrgDefaults();
   const t: IBurritoStrings = useSelector(burritoSelector, shallowEqual);
   const ts: ISharedStrings = useSelector(sharedSelector, shallowEqual);
@@ -67,7 +66,7 @@ export function BurritoWrapper() {
     if (teamId) {
       if (teamId) {
         const curContents = getOrgDefault(burritoWrapper, teamId) as
-          | BurritoWrapperType
+          | BurritoWrapper
           | undefined;
         if (curContents) {
           setMetaData(curContents);
@@ -93,7 +92,7 @@ export function BurritoWrapper() {
                     id: `${abbreviation}-${toCamel(c)}`,
                     path: toCamel(c).toLocaleLowerCase(),
                     role: burritoRole(c as BurritoType),
-                  }) as Burrito
+                  }) as BurritoEntry
               ) || [];
 
             setMetaData(
