+++ src/renderer/src/burrito/data/burritoBuilder.ts
@@ -1,154 +1,14 @@
-interface BurritoMeta {
-  version: string;
-  category: string;
-  generator: {
-    softwareName: string;
-    softwareVersion: string;
-    userName: string;
-  };
-  defaultLocale: string;
-  dateCreated: string;
-  comments: string[];
-}
-
-interface BurritoIdentification {
-  primary?: {
-    [key: string]: {
-      [key: string]: {
-        revision: string;
-        timestamp: string;
-      };
-    };
-  };
-  name?: {
-    [key: string]: string;
-  };
-  description?: {
-    [key: string]: string;
-  };
-  abbreviation?: {
-    [key: string]: string;
-  };
-}
-
-interface BurritoLanguage {
-  tag: string;
-  name: {
-    [key: string]: string;
-  };
-}
-
-// interface BurritoFormat {
-//   compression: string;
-//   trackConfiguration?: string;
-//   bitRate?: number;
-//   bitDepth?: number;
-//   samplingRate?: number;
-// }
-
-// export interface BurritoFormats {
-//   [key: string]: BurritoFormat;
-// }
-
-interface BurritoFlavor {
-  name: string;
-  // performance: string[];
-  // formats: BurritoFormats;
-}
-
-export interface BurritoScopes {
-  [book: string]: string[];
-}
-
-interface BurritoType {
-  flavorType: {
-    name: string;
-    flavor: BurritoFlavor;
-    currentScope: BurritoScopes;
-  };
-}
-
-interface BurritoAgency {
-  id: string;
-  roles: string[];
-  url?: string;
-  name: {
-    [key: string]: string;
-  };
-  abbr?: {
-    [key: string]: string;
-  };
-}
-
-interface BurritoTargetArea {
-  code: string;
-  name: {
-    [key: string]: string;
-  };
-}
-
-interface BurritoLocalizedName {
-  abbr: {
-    [key: string]: string;
-  };
-  short: {
-    [key: string]: string;
-  };
-  long: {
-    [key: string]: string;
-  };
-}
-
-export interface BurritoLocalizedNames {
-  [key: string]: BurritoLocalizedName;
-}
-
-interface BurritoIngredient {
-  checksum: {
-    md5: string;
-  };
-  mimeType: string;
-  size?: number;
-  scope?: {
-    [key: string]: string[];
-  };
-  role?: string;
-  properties?: {
-    [key: string]: string;
-  };
-}
-
-export interface BurritoIngredients {
-  [key: string]: BurritoIngredient;
-}
-
-interface BurritoCopyright {
-  shortStatements: {
-    statement: string;
-    mimetype: string;
-    lang: string;
-  }[];
-}
-
-export interface Burrito {
-  format: string;
-  meta: BurritoMeta;
-  idAuthorities?: {
-    [key: string]: {
-      id: string;
-      name: { [key: string]: string };
-    };
-  };
-  identification?: BurritoIdentification;
-  languages?: BurritoLanguage[];
-  type?: BurritoType;
-  confidential?: boolean;
-  agencies?: BurritoAgency[];
-  targetAreas?: BurritoTargetArea[];
-  localizedNames?: BurritoLocalizedNames;
-  ingredients: BurritoIngredients;
-  copyright?: BurritoCopyright;
-}
+import {
+  Burrito,
+  BurritoAgency,
+  BurritoCopyright,
+  BurritoFlavor,
+  BurritoIdentification,
+  BurritoIngredient,
+  BurritoLocalizedName,
+  BurritoLocalizedNames,
+  BurritoMeta,
+} from 'burrito/data/types';
 
 export class BurritoBuilder {
   private burrito: Burrito = {
