+++ src/renderer/src/burrito/data/wrapperBuilder.ts
@@ -1,17 +1,10 @@
-export interface Burrito {
-  id: string;
-  path: string;
-  role: string;
-}
-
-interface Alignment {
-  source: string;
-  target: string;
-  path: string;
-  description: {
-    en: string;
-  };
-}
+import {
+  Alignment,
+  BurritoEntry,
+  BurritoWrapper,
+  Contents,
+  WrapperMeta,
+} from 'burrito/data/types';
 
 interface RequiredParams {
   name: string;
@@ -25,41 +18,11 @@ interface OptionalParams {
   comments?: string;
   genName?: string;
   genVersion?: string;
-  burritos?: Burrito[];
+  burritos?: BurritoEntry[];
   alignments?: Alignment[];
 }
 
 type WrapperParams = RequiredParams & OptionalParams;
-
-interface Meta {
-  version: string;
-  name: {
-    en: string;
-  };
-  abbreviation?: {
-    en: string;
-  };
-  description: {
-    en: string;
-  };
-  generator?: {
-    name: string;
-    version?: string;
-  };
-  dateCreated?: string;
-  comments?: string;
-}
-
-interface Contents {
-  burritos: Burrito[];
-  alignments?: Alignment[];
-}
-
-export interface BurritoWrapper {
-  format: string;
-  meta: Meta;
-  contents: Contents;
-}
 
 /**
  * Creates a burrito wrapper object with the specified parameters
@@ -83,18 +46,20 @@ export function wrapperBuilder({
     name: {
       en: name,
     },
-  } as Meta;
+    description: {
+      en: description,
+    },
+    dateCreated: dateCreated || new Date().toISOString().split('T')[0],
+  } as WrapperMeta;
   if (abbreviation) {
     meta.abbreviation = { en: abbreviation };
   }
-  meta.description = { en: description };
   if (genName) {
     meta.generator = { name: genName };
     if (genVersion) {
       meta.generator.version = genVersion;
     }
   }
-  meta.dateCreated = dateCreated || new Date().toISOString().split('T')[0];
   if (comments) {
     meta.comments = comments;
   }
